testfix

+++ grammar/C/test/w_optimalisation5.c
@@ -1,7 +1,7 @@
 int main (){
 
-    float c = 1.0;
-    c = c * 1.0;
+    int c = 1;
+    c = c * 1;
 
 
     return 1;
